please stop deleting your own messages

+++ src/cogs/monkey_guild.py
@@ -134,7 +134,7 @@ class Monkey(commands.Cog):
 
     @commands.Cog.listener()
     async def on_message(self, message: discord.Message):
-        if message.author.id == self.bot.user.id:
+        if message.author.id == self.bot.user.id and message.channel.id == config.counting_channel_id:
             await asyncio.sleep(10)
             try:
                 await message.delete()
